Refactored python module structure
