+++ ddr_server/ddb_util.py
